requirements.txt
aiohttp>=3.9,<4
gigachat>=0.1.0
maxapi>=0.9,<1
pydantic>=2.0,<3
python-dotenv>=1.0,<2
